Watchlist y tendencias

+++ filmss/database/migrations/2023_05_10_164959_create_watchlists_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
             $table->string('contenido');
-            $table->float('estado');
+            $table->integer('estado');
             $table->timestamps();
 
             $table->primary(['user_id', 'contenido']);
